Nengo-wide default dtype can be set in RC
This makes it possible to set a default NumPy dtype that will be
used in all the places that we previously used 'float64'.

In doing some limited testing, using 'float32' results in little
to no speed difference, but a large reduction in memory usage.
'float16' reduces memory usage even more, but runs into floating
point math errors in many unit tests.

+++ nengo/spa/pointer.py
@@ -1,6 +1,7 @@
 import numpy as np
 
 from nengo.exceptions import ValidationError
+from nengo.rc import global_dtype
 from nengo.utils.numpy import is_integer, is_number
 
 
@@ -25,7 +26,7 @@ class SemanticPointer:
                 raise ValidationError(
                     "Must specify either the data or the length for a "
                     "SemanticPointer.", attr='data', obj=self)
-            self.v = np.array(data, dtype=float)
+            self.v = np.array(data, dtype=global_dtype)
             if len(self.v.shape) != 1:
                 raise ValidationError("'data' must be a vector", 'data', self)
